requirement.txt
datasets==2.19.0
diffusers==0.31.0
einops==0.8.0
huggingface-hub==0.30.2
imageio==2.34.1
jaxtyping==0.2.28
joblib==1.4.0
lightning-utilities==0.11.2
matplotlib==3.8.4
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
pandas==2.2.2
pillow==10.3.0
plyfile==1.0.3
PyMCubes==0.1.4
pyparsing==3.1.2
pytorch-lightning==2.2.4
torchmetrics==1.4.0.post0
PyYAML==6.0.1
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
tensorboard==2.16.2
tensorboardX==2.6.2.2
timm==0.9.16
tokenizers==0.21.1
tqdm==4.66.2
transformers==4.51.1
trimesh==4.3.2
spaces==0.28.3
accelerate==0.29.1
rembg==2.0.59
gradio==5.5.0
wandb==0.18.6
open3d
easydict
xformers==0.0.29.post1
pytorch_msssim
lpips
imageio-ffmpeg
kiui
pymeshlab==2023.12
